Modificaciones en el juego, se agrego una entrada para el user
